requirements.txt
attrs==17.4.0
easydict==1.7
falcon==1.4.1
gunicorn==19.7.1
Jinja2==2.10
MarkupSafe==1.0
python-mimeparse==1.6.0
six==1.11.0
toml==0.9.4
